backend/pyproject.toml
[project]
name = "stock-simulator-backend"
version = "1.0.0"
description = "KOSPI + NASDAQ algorithmic trading simulator - Django backend"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<6.0",
    "django-ninja>=1.3,<2.0",
    "django-cors-headers>=4.4",
    "finance-datareader",
    "pandas",
    "numpy",
    "pyarrow",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
